Added Hacking Two and updated some balancing issues.

+++ Models/Upgrades/Upgrade.cs
@@ -4,8 +4,8 @@
     {
         public bool Active;
         public bool Unlocked;
-        public float Multiplier { get; set; }
-        public float Expense { get; set; }
+        public decimal Multiplier { get; set; }
+        public decimal Expense { get; set; }
         public Modifiers Modifier { get; set; }
         public string Name { get; set; } = "";
         public UpgradeTypes UpgradeType { get; set; }
